# Prevent creating Point slices with a cap larger than the remaining samples