<!--
    Don't forget!
    
    - If the PR adds or changes a behaviour or fixes a bug of an exported API it would need a unit/e2e test.
    
    - Where possible use only exported APIs for tests to simplify the review and make it as close as possible to an actual library usage.
    
    - No tests are needed for internal implementation changes.
    
    - Performance improvements would need a benchmark test to prove it.
    
    - All exposed objects should have a comment.
    
    - All comments should start with a capital letter and end with a full stop.
 -->

Add a testcase with a large subquery and run the benchmark next to each other ten times. Benchstat seems to be happy with slightly less allocations after the change.


```
:~/environment $ ~/go/bin/benchstat  -sort delta ./github.com/prometheus/prometheus.old/before9 ./github.com/prometheus/prometheus/after9
name                                                                                                      old time/op    new time/op    delta
RangeQuery/expr=a_one,steps=1-2                                                                              147µs ±98%    119µs ±147%    ~     (p=0.604 n=9+10)
RangeQuery/expr=a_one,steps=10-2                                                                             162µs ±79%    124µs ±162%    ~     (p=0.218 n=10+10)
RangeQuery/expr=a_one,steps=100-2                                                                            199µs ±80%    152µs ±156%    ~     (p=0.684 n=10+10)
RangeQuery/expr=a_one,steps=1000-2                                                                          506µs ±117%    364µs ±134%    ~     (p=0.579 n=10+10)
RangeQuery/expr=a_ten,steps=1-2                                                                              422µs ±93%    308µs ±143%    ~     (p=0.143 n=10+10)
RangeQuery/expr=a_ten,steps=10-2                                                                             434µs ±76%    346µs ±124%    ~     (p=0.436 n=10+10)
RangeQuery/expr=a_ten,steps=100-2                                                                           858µs ±130%    716µs ±122%    ~     (p=0.912 n=10+10)
RangeQuery/expr=a_ten,steps=1000-2                                                                          3.59ms ±73%   3.16ms ±122%    ~     (p=0.853 n=10+10)
RangeQuery/expr=a_hundred,steps=1-2                                                                        2.53ms ±111%   2.59ms ±113%    ~     (p=0.971 n=10+10)
RangeQuery/expr=a_hundred,steps=10-2                                                                       2.86ms ±109%   2.76ms ±111%    ~     (p=0.579 n=10+10)
RangeQuery/expr=a_hundred,steps=100-2                                                                      7.02ms ±191%   7.05ms ±132%    ~     (p=0.971 n=10+10)
RangeQuery/expr=a_hundred,steps=1000-2                                                                     31.2ms ±101%   31.1ms ±116%    ~     (p=0.796 n=10+10)
RangeQuery/expr=rate(a_one[1m]),steps=1-2                                                                   162µs ±105%    163µs ±111%    ~     (p=0.579 n=10+10)
RangeQuery/expr=rate(a_one[1m]),steps=10-2                                                                  167µs ±100%    167µs ±121%    ~     (p=0.853 n=10+10)
RangeQuery/expr=rate(a_one[1m]),steps=100-2                                                                 242µs ±116%    245µs ±148%    ~     (p=0.912 n=10+10)
RangeQuery/expr=rate(a_one[1m]),steps=1000-2                                                                797µs ±120%    824µs ±140%    ~     (p=0.971 n=10+10)
RangeQuery/expr=rate(a_ten[1m]),steps=1-2                                                                   368µs ±107%    428µs ±148%    ~     (p=0.684 n=10+10)
RangeQuery/expr=rate(a_ten[1m]),steps=10-2                                                                  420µs ±104%    469µs ±111%    ~     (p=0.912 n=10+10)
RangeQuery/expr=rate(a_ten[1m]),steps=100-2                                                                1.16ms ±146%   1.15ms ±118%    ~     (p=0.853 n=10+10)
RangeQuery/expr=rate(a_ten[1m]),steps=1000-2                                                               5.44ms ±143%   5.99ms ±166%    ~     (p=0.971 n=10+10)
RangeQuery/expr=rate(a_hundred[1m]),steps=1-2                                                              2.12ms ±136%   2.49ms ±105%    ~     (p=0.393 n=10+10)
RangeQuery/expr=rate(a_hundred[1m]),steps=10-2                                                             2.65ms ±151%   3.14ms ±112%    ~     (p=0.796 n=10+10)
RangeQuery/expr=rate(a_hundred[1m]),steps=100-2                                                            8.36ms ±137%  10.29ms ±114%    ~     (p=0.912 n=10+10)
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-2                                                           56.4ms ±152%   71.2ms ±174%    ~     (p=0.912 n=10+10)
RangeQuery/expr=rate(a_one[1m]),steps=10000-2                                                              5.83ms ±136%   7.52ms ±147%    ~     (p=0.971 n=10+10)
RangeQuery/expr=rate(a_ten[1m]),steps=10000-2                                                              54.8ms ±133%   74.1ms ±136%    ~     (p=0.280 n=10+10)
RangeQuery/expr=rate(a_hundred[1m]),steps=10000-2                                                           542ms ±132%    735ms ±161%    ~     (p=0.143 n=10+10)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1-2                                                4.34ms ±135%   5.76ms ±163%    ~     (p=0.796 n=10+10)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=10-2                                               5.77ms ±132%   7.48ms ±171%    ~     (p=1.000 n=10+10)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=100-2                                              19.8ms ±145%   25.4ms ±171%    ~     (p=0.579 n=10+10)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1000-2                                              158ms ±124%    198ms ±165%    ~     (p=0.684 n=10+10)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1-2                                                26.5ms ±149%   35.8ms ±244%    ~     (p=0.971 n=10+10)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=10-2                                               40.0ms ±127%   50.4ms ±148%    ~     (p=0.631 n=10+10)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=100-2                                               209ms ±124%    243ms ±161%    ~     (p=0.739 n=10+10)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1000-2                                              1.75s ±101%    1.80s ±103%    ~     (p=0.579 n=10+10)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1-2                                              260ms ±94%    267ms ±103%    ~     (p=0.796 n=10+10)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=10-2                                            436ms ±109%    428ms ±100%    ~     (p=0.796 n=10+10)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-2                                           2.01s ±110%    2.00s ±100%    ~     (p=0.912 n=10+10)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1000-2                                          15.6s ±118%    16.1s ±115%    ~     (p=0.739 n=10+10)
RangeQuery/expr=changes(a_one[1d]),steps=1-2                                                               3.84ms ±108%   3.79ms ±107%    ~     (p=0.315 n=10+10)
RangeQuery/expr=changes(a_one[1d]),steps=10-2                                                              4.66ms ±104%   4.59ms ±133%    ~     (p=0.631 n=10+10)
RangeQuery/expr=changes(a_one[1d]),steps=100-2                                                             11.4ms ±100%   11.6ms ±101%    ~     (p=0.853 n=10+10)
RangeQuery/expr=changes(a_one[1d]),steps=1000-2                                                             78.3ms ±99%   81.7ms ±107%    ~     (p=0.796 n=10+10)
RangeQuery/expr=changes(a_ten[1d]),steps=1-2                                                               27.4ms ±127%   25.6ms ±105%    ~     (p=0.393 n=10+10)
RangeQuery/expr=changes(a_ten[1d]),steps=10-2                                                               31.8ms ±96%   32.8ms ±113%    ~     (p=0.853 n=10+10)
RangeQuery/expr=changes(a_ten[1d]),steps=100-2                                                              104ms ±100%     101ms ±99%    ~     (p=0.481 n=10+10)
RangeQuery/expr=changes(a_ten[1d]),steps=1000-2                                                             793ms ±106%     781ms ±97%    ~     (p=0.912 n=10+10)
RangeQuery/expr=changes(a_hundred[1d]),steps=1-2                                                            246ms ±110%    261ms ±164%    ~     (p=0.971 n=10+10)
RangeQuery/expr=changes(a_hundred[1d]),steps=10-2                                                           321ms ±115%     310ms ±98%    ~     (p=0.853 n=10+10)
RangeQuery/expr=changes(a_hundred[1d]),steps=100-2                                                           1.01s ±99%     0.99s ±96%    ~     (p=0.684 n=10+10)
RangeQuery/expr=changes(a_hundred[1d]),steps=1000-2                                                         7.51s ±106%    7.10s ±113%    ~     (p=1.000 n=10+10)
RangeQuery/expr=rate(a_one[1d]),steps=1-2                                                                  3.12ms ±146%   2.87ms ±131%    ~     (p=0.218 n=10+10)
RangeQuery/expr=rate(a_one[1d]),steps=10-2                                                                 3.48ms ±149%   3.25ms ±136%    ~     (p=0.853 n=10+10)
RangeQuery/expr=rate(a_one[1d]),steps=100-2                                                                6.55ms ±147%   6.58ms ±130%    ~     (p=0.853 n=10+10)
RangeQuery/expr=rate(a_one[1d]),steps=1000-2                                                               34.7ms ±135%   34.5ms ±138%    ~     (p=0.739 n=10+10)
RangeQuery/expr=rate(a_ten[1d]),steps=1-2                                                                  22.1ms ±140%   22.2ms ±164%    ~     (p=0.796 n=10+10)
RangeQuery/expr=rate(a_ten[1d]),steps=10-2                                                                 25.0ms ±141%   25.3ms ±151%    ~     (p=0.579 n=10+10)
RangeQuery/expr=rate(a_ten[1d]),steps=100-2                                                                57.3ms ±150%   58.2ms ±119%    ~     (p=0.143 n=10+10)
RangeQuery/expr=rate(a_ten[1d]),steps=1000-2                                                                344ms ±123%    357ms ±131%    ~     (p=0.579 n=10+10)
RangeQuery/expr=rate(a_hundred[1d]),steps=1-2                                                               217ms ±121%    227ms ±127%    ~     (p=0.739 n=10+10)
RangeQuery/expr=rate(a_hundred[1d]),steps=10-2                                                              287ms ±123%     272ms ±98%    ~     (p=0.739 n=10+10)
RangeQuery/expr=rate(a_hundred[1d]),steps=100-2                                                             637ms ±121%    660ms ±113%    ~     (p=0.123 n=10+10)
RangeQuery/expr=rate(a_hundred[1d]),steps=1000-2                                                            3.92s ±102%     3.88s ±96%    ~     (p=0.579 n=10+10)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=1-2                                                      3.69ms ±109%   3.59ms ±108%    ~     (p=0.912 n=10+10)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=10-2                                                     3.81ms ±102%   3.72ms ±103%    ~     (p=0.315 n=10+10)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=100-2                                                    5.44ms ±105%   5.27ms ±109%    ~     (p=0.796 n=10+10)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=1000-2                                                   19.4ms ±105%   19.1ms ±100%    ~     (p=0.912 n=10+10)
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=1-2                                                      26.2ms ±127%    24.5ms ±98%    ~     (p=0.481 n=10+10)
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=10-2                                                     26.4ms ±105%   25.4ms ±107%    ~     (p=0.393 n=10+10)
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=100-2                                                    42.3ms ±115%   33.4ms ±128%    ~     (p=0.190 n=10+10)
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=1000-2                                                     183ms ±98%    154ms ±133%    ~     (p=1.000 n=10+10)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1-2                                                   243ms ±115%    195ms ±132%    ~     (p=0.631 n=10+10)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=10-2                                                  258ms ±139%    218ms ±157%    ~     (p=0.631 n=10+10)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=100-2                                                  408ms ±98%    342ms ±126%    ~     (p=0.631 n=10+10)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1000-2                                                 1.84s ±92%    1.60s ±126%    ~     (p=0.796 n=10+10)
RangeQuery/expr=-a_one,steps=1-2                                                                             158µs ±77%    141µs ±102%    ~     (p=1.000 n=10+10)
RangeQuery/expr=-a_one,steps=10-2                                                                            172µs ±87%     146µs ±99%    ~     (p=0.315 n=10+10)
RangeQuery/expr=-a_one,steps=100-2                                                                           211µs ±75%    177µs ±107%    ~     (p=0.579 n=10+10)
RangeQuery/expr=-a_one,steps=1000-2                                                                         529µs ±145%    420µs ±111%    ~     (p=0.971 n=10+10)
RangeQuery/expr=-a_ten,steps=1-2                                                                             442µs ±94%     358µs ±98%    ~     (p=0.280 n=10+10)
RangeQuery/expr=-a_ten,steps=10-2                                                                            478µs ±86%    389µs ±106%    ~     (p=0.190 n=10+10)
RangeQuery/expr=-a_ten,steps=100-2                                                                          925µs ±126%    719µs ±111%    ~     (p=0.123 n=10+10)
RangeQuery/expr=-a_ten,steps=1000-2                                                                         3.73ms ±81%   3.12ms ±100%    ~     (p=0.739 n=10+10)
RangeQuery/expr=-a_hundred,steps=1-2                                                                        3.17ms ±80%   2.80ms ±115%    ~     (p=0.912 n=10+10)
RangeQuery/expr=-a_hundred,steps=10-2                                                                       3.58ms ±81%    2.93ms ±97%    ~     (p=0.481 n=10+10)
RangeQuery/expr=-a_hundred,steps=100-2                                                                     7.70ms ±104%   6.42ms ±112%    ~     (p=0.481 n=10+10)
RangeQuery/expr=-a_hundred,steps=1000-2                                                                    41.6ms ±135%   32.0ms ±104%    ~     (p=0.353 n=10+10)
RangeQuery/expr=a_one_-_b_one,steps=1-2                                                                      242µs ±79%    205µs ±111%    ~     (p=0.796 n=10+10)
RangeQuery/expr=a_one_-_b_one,steps=10-2                                                                     277µs ±80%     228µs ±98%    ~     (p=0.481 n=10+10)
RangeQuery/expr=a_one_-_b_one,steps=100-2                                                                    697µs ±89%    603µs ±110%    ~     (p=0.353 n=10+10)
RangeQuery/expr=a_one_-_b_one,steps=1000-2                                                                 3.61ms ±106%    3.61ms ±97%    ~     (p=0.684 n=10+10)
RangeQuery/expr=a_ten_-_b_ten,steps=1-2                                                                     776µs ±109%    790µs ±103%    ~     (p=0.631 n=10+10)
RangeQuery/expr=a_ten_-_b_ten,steps=10-2                                                                   1.10ms ±110%   1.13ms ±109%    ~     (p=0.393 n=10+10)
RangeQuery/expr=a_ten_-_b_ten,steps=100-2                                                                  4.65ms ±136%   4.48ms ±112%    ~     (p=0.739 n=10+10)
RangeQuery/expr=a_ten_-_b_ten,steps=1000-2                                                                 35.5ms ±121%   36.0ms ±109%    ~     (p=0.912 n=10+10)
RangeQuery/expr=a_hundred_-_b_hundred,steps=1-2                                                            6.76ms ±123%    7.83ms ±84%    ~     (p=1.000 n=10+10)
RangeQuery/expr=a_hundred_-_b_hundred,steps=10-2                                                           10.3ms ±104%    11.6ms ±75%    ~     (p=0.631 n=10+10)
RangeQuery/expr=a_hundred_-_b_hundred,steps=100-2                                                          45.8ms ±108%   59.2ms ±155%    ~     (p=0.796 n=10+10)
RangeQuery/expr=a_hundred_-_b_hundred,steps=1000-2                                                          373ms ±108%     425ms ±84%    ~     (p=0.739 n=10+10)
RangeQuery/expr=a_one_-_b_one,steps=10000-2                                                                35.3ms ±103%    40.9ms ±77%    ~     (p=0.579 n=10+10)
RangeQuery/expr=a_ten_-_b_ten,steps=10000-2                                                                 365ms ±174%     392ms ±77%    ~     (p=1.000 n=10+10)
RangeQuery/expr=a_hundred_-_b_hundred,steps=10000-2                                                         3.88s ±107%     4.46s ±79%    ~     (p=0.912 n=10+10)
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1-2                                                    395µs ±122%     460µs ±81%    ~     (p=0.796 n=10+10)
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=10-2                                                   413µs ±122%     449µs ±73%    ~     (p=0.739 n=10+10)
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=100-2                                                  635µs ±116%     672µs ±82%    ~     (p=0.631 n=10+10)
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1000-2                                                2.15ms ±105%    2.51ms ±79%    ~     (p=1.000 n=10+10)
RangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1-2                                                    811µs ±105%     921µs ±91%    ~     (p=0.631 n=10+10)
RangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=10-2                                                   968µs ±118%    1090µs ±92%    ~     (p=0.912 n=10+10)
RangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=100-2                                                 2.75ms ±156%   3.26ms ±132%    ~     (p=0.912 n=10+10)
RangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1000-2                                                16.8ms ±104%   22.3ms ±100%    ~     (p=0.481 n=10+10)
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1-2                                           5.17ms ±103%    5.51ms ±74%    ~     (p=0.436 n=10+10)
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=10-2                                          7.08ms ±119%   8.17ms ±111%    ~     (p=0.971 n=10+10)
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=100-2                                         29.7ms ±159%    29.2ms ±77%    ~     (p=0.315 n=10+10)
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1000-2                                         202ms ±105%    235ms ±108%    ~     (p=0.739 n=10+10)
RangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1-2                                                     429µs ±108%     469µs ±82%    ~     (p=0.853 n=10+10)
RangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=10-2                                                    427µs ±109%     476µs ±91%    ~     (p=1.000 n=10+10)
…