@roidelapluie 2023-09-05
cc @bboreham what do you think of this change?

@zecke do you still want to work on this and rebase this?