@alanprot 2023-09-05
Oh.. i missed this one..

This seems similar to the problem we saw as well and tried to fix on https://github.com/prometheus/prometheus/pull/12734

The problem of this approach is that we can still go OOM for sparse series as discussed here: https://github.com/prometheus/prometheus/pull/12734/files#r1306045499

In summary: We are allocating always the `min(numSteps, ev.maxSamples-ev.currentSamples)` but if the result does not brings lots of points the `ev.currentSamples` will still be low and so, we will keep allocating more memory.

I think the conclusion there is to allocate slices of max 5k or 10k (`min(numSteps, 5k)`). WDYT?