@zecke 2023-09-12
I probably won't have cycles to rebase this and see it to completion (brian successfully blocked it). I think this change and #7930 were mitigations from an outage. We don't want to over allocate and then hold onto the memory (like putting it back to a sync.Pool). The theory is kind of that the cost of growing the slice is in the noise for larger queries anyway.

With more time I would have attempted to get something that mostly behaves like a slice that is backed by 0..n fixed size slices (think of a slab allocator..)