@zecke 2023-10-05
> This was done in #12734.

Thanks for the pointer and great to see this land. What do you think about putting slices with a high capacity into the sync pool. My original patch did this: https://github.com/prometheus/prometheus/pull/7930/files#diff-ad2761bb162a2f12326b18f2714ef54f4cd476cfe3388f77b2b20c5ff82e4faaR1422-R1431