@bboreham 2023-10-05
> What do you think about putting slices with a high capacity into the sync pool.

Looking at the code I think you mean "_not_ putting slices".

Did you measure the impact of this?  

I wouldn't expect it makes much difference as (a) unused pool elements are garbage-collected eventually and (b) if we do actually create slices that large then we will likely need them again, in which case keeping them in the pool is valuable.
The basis of #12734 was that we likely never really need large slices, so we will never create them.