@zecke 2023-10-06
> > What do you think about putting slices with a high capacity into the sync pool.
> 
> Looking at the code I think you mean "_not_ putting slices".
> 
> Did you measure the impact of this?
> 
> I wouldn't expect it makes much difference as (a) unused pool elements are garbage-collected eventually and (b) if we do actually create slices that large then we will likely need them again, in which case keeping them in the pool is valuable. The basis of #12734 was that we likely never really need large slices, so we will never create them.

Oops. Yeah: "_not_ putting a large cap slice into the pool". I don't remember to which degree I benchmarked this but IIRC the line of thought is something like:

1.) If (b) is true then we will have _multiple_ large cap slices in the pool (as the pool is not a slab allocator and there is no guarantee that we get back the same slice for the same rule)
2.) large cap slices are directly allocated from the heap

=> 1.) + 2.) we ended up holding on to a lot more memory than needed, and potentially fragment the heap.