@bboreham 2023-10-06
OK, please update if you have data.  
There is a pool for multiple bucket sizes in `prometheus/util/pool`.