@beorn7 2024-02-01
:wave: from the bug scrub. With no movement on this in a while, we'll close. Note that there is #13448 , which seems to solves the same problem.